RAGchatbot/requirements.txt
langchain
langchain-core
langchain-community
python-dotenv
Ollama
langchain-groq

sentence-transformers